football_dashboard_backend/requirements.txt
Flask
flask-cors
psycopg2-binary
python-dotenv
google-genai==1.7.0
pandas
requests
